requirements.txt
streamlit
openai>=1.0.0
python-dotenv
pandas
pytest
pytest-mock
requests

# API server
fastapi
uvicorn[standard]
pydantic
